refactor: move dialogs to its own components
