+++ src/modules/dapp/pages/index.ts
@@ -1,3 +1,4 @@
 export * from './transactionConfirm';
 //export * from './transactionSign'; [CONNECTOR SIGNATURE]
 export * from './VaultConnector';
+export * from './welcome';
